+++ livnote/src-tauri/src/listners/mod.rs
@@ -1,0 +1,90 @@
+use std::sync::Arc;
+
+use crate::{current_note_state::CurrentNoteState, user_state::UserState};
+use crypto_utils::CryptoUtils;
+use log::info;
+use network::p2p::{P2PEvent, incoming::P2PSender};
+use persistance::database::RepositoryContext;
+use serde::{Deserialize, Serialize};
+use tauri::{AppHandle, Manager};
+use tokio::sync::{Mutex, mpsc};
+
+mod p2p_handlers;
+mod p2p_reciever;
+mod p2p_sender;
+mod tauri_events;
+
+/// Initializes all listeners for the application
+/// This connects the Tauri event system with the P2P event system
+pub struct EventManager {
+    app_handle: AppHandle,
+    p2p_receiver: mpsc::UnboundedReceiver<P2PEvent>,
+    p2p_sender: P2PSender,
+    current_note_state: CurrentNoteState,
+    repo_ctx: Arc<RepositoryContext>,
+    crypto_utils: Arc<Mutex<CryptoUtils>>,
+}
+
+#[derive(Debug, Clone)]
+enum UpdateType {
+    SyncUpdate,
+    AwarenessUpdate,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct NoteChangePayload {
+    pub note_id: Option<String>,
+    pub state_vectors: serde_json::Value,
+}
+
+impl UpdateType {
+    fn event_name(&self) -> &'static str {
+        match self {
+            UpdateType::SyncUpdate => "sync-update",
+            UpdateType::AwarenessUpdate => "awareness-update",
+        }
+    }
+
+    fn description(&self) -> &'static str {
+        match self {
+            UpdateType::SyncUpdate => "sync update",
+            UpdateType::AwarenessUpdate => "awareness update",
+        }
+    }
+}
+
+impl EventManager {
+    /// Create a new EventManager that handles bidirectional events
+    pub fn new(
+        app_handle: AppHandle,
+        p2p_receiver: mpsc::UnboundedReceiver<P2PEvent>,
+        p2p_sender: P2PSender,
+        repo_ctx: Arc<RepositoryContext>,
+        crypto_utils: Arc<Mutex<CryptoUtils>>,
+    ) -> Self {
+        Self {
+            app_handle,
+            p2p_receiver,
+            p2p_sender,
+            current_note_state: CurrentNoteState::new(),
+            repo_ctx,
+            crypto_utils,
+        }
+    }
+
+    /// Start listening for all events
+    pub fn start_listening(mut self) {
+        // Set up Tauri event listeners
+        self.setup_tauri_listeners();
+
+        // Start P2P event listener in background
+        tokio::spawn(async move {
+            self.listen_for_p2p_events().await;
+        });
+    }
+
+    /// Add a method to access the current note state
+    pub fn get_current_note_state(&self) -> CurrentNoteState {
+        self.current_note_state.clone()
+    }
+}
