+++ livnote/src-tauri/src/listners/p2p_handlers/system.rs
@@ -1,0 +1,53 @@
+use crate::EventManager;
+use log::{error, info};
+use tauri::Emitter;
+
+/// System-level P2P event handlers (connection lifecycle, sync status, errors)
+impl EventManager {
+    /// Handle connected event
+    pub(crate) fn handle_connected_event(&self) {
+        if let Err(e) = self.app_handle.emit("peer-connected", true) {
+            error!("Failed to emit peer-connected event: {}", e);
+        }
+    }
+
+    /// Handle disconnected event
+    pub(crate) fn handle_disconnected_event(&self) {
+        info!("Peer disconnected");
+        if let Err(e) = self.app_handle.emit("peer-disconnected", true) {
+            error!("Failed to emit peer-disconnected event: {}", e);
+        }
+    }
+
+    /// Handle handshake failed event
+    pub(crate) fn handle_handshake_failed_event(&self, error: String) {
+        info!("Handshake failed: {}", error);
+        if let Err(e) = self.app_handle.emit("handshake-failed", error) {
+            error!("Failed to emit handshake-failed event: {}", e);
+        }
+    }
+
+    /// Handle sync complete event
+    pub(crate) fn handle_sync_complete_event(&self) {
+        info!("Sync completed");
+        if let Err(e) = self.app_handle.emit("sync-complete", true) {
+            error!("Failed to emit sync-complete event: {}", e);
+        }
+    }
+
+    /// Handle share complete event
+    pub(crate) fn handle_share_complete_event(&self) {
+        info!("Share completed");
+        if let Err(e) = self.app_handle.emit("share-complete", true) {
+            error!("Failed to emit share-complete event: {}", e);
+        }
+    }
+
+    /// Handle error event
+    pub(crate) fn handle_error_event(&self, message: String, source: String) {
+        error!("P2P error from {}: {}", source, message);
+        if let Err(e) = self.app_handle.emit("p2p-error", message) {
+            error!("Failed to emit p2p-error event: {}", e);
+        }
+    }
+}
